explore fixed bug

+++ src/app/explore/explore.component.ts
@@ -16,6 +16,12 @@ export class ExploreComponent implements OnInit {
   isFormValid: boolean = false;
   conflict: boolean = false;
   isErrorOccurred:boolean = false;
+  alertCantDeleted: string = '';
+  createMixName: string = '';
+  creator: string = '';
+  trackId1:number;
+  trackId2:number;
+  trackId3:number;
 
   constructor(private apiService: ApiService) { }
 
@@ -36,14 +42,29 @@ export class ExploreComponent implements OnInit {
   }
 
   onDelete(val) {
+    if(!val.deleteMixName){
+      val.deleteMixName = 'first mix';
+      this.alertCantDeleted=`Failure: <i>${val.deleteMixName}</i> was not Deleted, you can't delete mix from 1 to 15`;
+      return;
+    }
+    if(val.deleteMixName === 'first mix' || val.deleteMixName === 'second mix' ||val.deleteMixName === 'third mix' ||val.deleteMixName === 'fourth mix' ||
+      val.deleteMixName === 'fifth mix' ||val.deleteMixName === 'sixth mix' ||val.deleteMixName === 'seventh mix' ||val.deleteMixName === 'eighth mix' ||
+      val.deleteMixName === 'ninth mix' ||val.deleteMixName === 'tenth mix' ||val.deleteMixName === 'eleventh mix' ||val.deleteMixName === 'twelfth mix' ||
+      val.deleteMixName === 'thirteen mix' ||val.deleteMixName === 'fourteen mix' ||val.deleteMixName === 'fifteen mix'){
+      this.alertDeleted='';
+      this.alertCantDeleted=`Failure: <i>${val.deleteMixName}</i> was not Deleted, you can't delete mix from 1 to 15`;
+      return;
+    }
     this.apiService.deleteMix(val.deleteMixName).then(() => {this.isErrorOccurred = false;})
       .catch(err => {
         console.log(`error when trying to connect to server ${err}`);
         this.isErrorOccurred = true;
+        this.alertCantDeleted ='';
         this.alertDeleted=`Failure: <i>${val.deleteMixName}</i> was not Deleted`;
 
       });
     this.ngOnInit();
+    this.alertCantDeleted ='';
     this.alertDeleted=`Success: <i>${val.deleteMixName}</i> was Deleted`;
   }
   private createMix(name, creator, track1, track2, track3){
@@ -56,9 +77,9 @@ export class ExploreComponent implements OnInit {
     this.ngOnInit();
   }
 
-  onCreate(val){
+  onForm(val){
     this.checkFormValidity(val);
-
+    this.clear();
     if(this.isFormValid){
       this.createMix(val.createMixName,val.creator,val.trackId1,val.trackId2,val.trackId3);
       this.successMsg(val);
@@ -68,7 +89,13 @@ export class ExploreComponent implements OnInit {
       this.conflictTrack(val);
     }
   }
-
+  clear(){
+    this.createMixName = null;
+    this.creator = null;
+    this.trackId1 = null;
+    this.trackId2 = null;
+    this.trackId3 = null;
+  }
   checkFormValidity(val) {
     if(val.trackId1 === val.trackId2 || val.trackId1 === val.trackId3 || val.trackId2 === val.trackId3){
       this.isFormValid = false;
